64bit safe types.
git-svn-id: http://sasl.googlecode.com/svn/trunk@205 84095216-3abd-11de-90e1-f9f1724fc9cc

+++ xap/gui.cpp
@@ -44,7 +44,7 @@ static XPWidgetID okButton;
 
 // called on login dialog events
 static int optionsWindowHandler(XPWidgetMessage message, XPWidgetID widget,
-        long param1, long param2)
+        intptr_t param1, intptr_t param2)
 {
     switch (message) {
         case xpMessage_CloseButtonPushed:
